Rename 'components/SevereCaseBarChart.vue' to 'components/index/CardsMonitoring/SevereCase/Chart.vue'
